Added quote generation to focusbot
