+++ Motivational_Quotes/motivation.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-"""Motivation.ipynb
+"""
+Motivation.ipynb
 
 Automatically generated by Colaboratory.
 
@@ -10,17 +11,17 @@ Original file is located at
 import pandas as pd
 import random
 
+#Reads in a database of quotes
 df = pd.read_json("Quotes\quotes.json")
 
-print(df.head())
-
+#dropping the popularity axis 
 df.drop('Popularity', axis=1, inplace=True)
-print(df.head())
 
+#getting a list of categories
 cat = df['Category'].unique()
 print(cat)
 
-# #Depreciated :(
+# #Depreciated old way thats inefficient :(
 # filtered = pd.DataFrame()
 # for item in cat:
 #     if(item == 'inspiration'):
@@ -31,18 +32,15 @@ print(cat)
 #         filtered = pd.concat([df[df['Category']=='positive'],filtered])
 # print(len(filtered))
 
+#filtering on necessary categories
 filtered = df.loc[df['Category'].isin(['motivation','inspiration','positive'])]
 print(len(filtered))
 
+#removing the quotes that fell under multiple categories
 filtered = filtered.loc[filtered.astype(str).drop_duplicates().index]
 print(len(filtered))
 
-result= filtered.to_json(r"Quotes\filtered.json")
-
-test = pd.read_json("Quotes\\filtered.json")
-count = len(test)
-num = random.randint(0,count+1)
-print(test.iloc[num]['Quote'])
-
+#stroing the quotes in a new JSON
+result = filtered.to_json(r"Quotes\filtered.json")
